+++ tests/test_sim.py
@@ -15,10 +15,12 @@
 from io import BytesIO
 import gzip
 import os
-import os.path
 import stat
+from os.path import join
+from shutil import make_archive
 
 from mrjob import conf
+from mrjob.examples.mr_word_freq_count import MRWordFreqCount
 from mrjob.inline import InlineMRJobRunner
 from mrjob.job import MRJob
 from mrjob.protocol import JSONValueProtocol
@@ -26,6 +28,7 @@ from mrjob.step import MRStep
 
 from tests.mr_group import MRGroup
 from tests.mr_no_mapper import MRNoMapper
+from tests.mr_os_walk_job import MROSWalkJob
 from tests.mr_sort_and_group import MRSortAndGroup
 from tests.mr_test_jobconf import MRTestJobConf
 from tests.mr_test_per_step_jobconf import MRTestPerStepJobConf
@@ -122,7 +125,7 @@ class MRJobFileOptionsTestCase(SandboxedTestCase):
     def setUp(self):
         super(MRJobFileOptionsTestCase, self).setUp()
 
-        self.input_file_path = os.path.join(self.tmp_dir, 'input_file.txt')
+        self.input_file_path = join(self.tmp_dir, 'input_file.txt')
         with open(self.input_file_path, 'wb') as f:
             f.write(b'2\n')
 
@@ -181,11 +184,11 @@ class SimRunnerJobConfTestCase(SandboxedTestCase):
     RUNNER = 'inline'
 
     def test_input_files(self):
-        input_path = os.path.join(self.tmp_dir, 'input')
+        input_path = join(self.tmp_dir, 'input')
         with open(input_path, 'wb') as input_file:
             input_file.write(b'bar\nqux\nfoo\n')
 
-        input_gz_path = os.path.join(self.tmp_dir, 'input.gz')
+        input_gz_path = join(self.tmp_dir, 'input.gz')
         with gzip.GzipFile(input_gz_path, 'wb') as input_gz:
             input_gz.write(b'foo\n')
 
@@ -213,12 +216,12 @@ class SimRunnerJobConfTestCase(SandboxedTestCase):
 
     def test_jobconf_simulated_by_runner(self):
         # use a .gz file so there's only one split
-        input_gz_path = os.path.join(self.tmp_dir, 'input.gz')
+        input_gz_path = join(self.tmp_dir, 'input.gz')
         with gzip.GzipFile(input_gz_path, 'wb') as input_gz:
             input_gz.write(b'foo\n')
         input_gz_size = os.stat(input_gz_path)[stat.ST_SIZE]
 
-        upload_path = os.path.join(self.tmp_dir, 'upload')
+        upload_path = join(self.tmp_dir, 'upload')
         with open(upload_path, 'wb') as upload_file:
             upload_file.write(b'PAYLOAD')
 
@@ -248,8 +251,8 @@ class SimRunnerJobConfTestCase(SandboxedTestCase):
 
         working_dir = results['mapreduce.job.local.dir']
         self.assertEqual(working_dir,
-                         os.path.join(runner._get_local_tmp_dir(),
-                                      'step', '000', 'mapper', '00000', 'wd'))
+                         join(runner._get_local_tmp_dir(),
+                              'step', '000', 'mapper', '00000', 'wd'))
 
         self.assertEqual(results['mapreduce.job.cache.archives'], '')
 
@@ -266,10 +269,10 @@ class SimRunnerJobConfTestCase(SandboxedTestCase):
 
         self.assertEqual(results['mapreduce.job.cache.local.archives'], '')
         expected_local_files = [
-            os.path.join(working_dir, 'mr_test_jobconf.py'),
-            os.path.join(working_dir, 'upload')
+            join(working_dir, 'mr_test_jobconf.py'),
+            join(working_dir, 'upload')
         ] + [
-            os.path.join(working_dir, name)
+            join(working_dir, name)
             for path, name in self._extra_expected_local_files(runner)
         ]
         self.assertEqual(
@@ -339,9 +342,15 @@ class LocalFSTestCase(SandboxedTestCase):
 
     def test_can_handle_paths(self):
         self.assertEqual(
-            self.runner.fs.exists(os.path.join(self.tmp_dir, 'foo')), False)
+            self.runner.fs.exists(join(self.tmp_dir, 'foo')), False)
 
-    def test_cant_handle_uris(self):
+    def test_can_handle_file_uris(self):
+        self.assertEqual(
+            self.runner.fs.exists(
+                'file://' + join(self.tmp_dir, 'foo')),
+            False)
+
+    def test_cant_handle_other_uris(self):
         self.assertRaises(IOError, self.runner.fs.ls, 's3://walrus/foo')
 
 
@@ -367,20 +376,77 @@ class DistributedCachePermissionsTestCase(SandboxedTestCase):
         self.assertFalse(data_perms & stat.S_IXOTH)
 
 
-class FSDoesntHandleURIsTestCase(SandboxedTestCase):
+class FSOnlyHandlesFileURIsTestCase(SandboxedTestCase):
     # regression test for #1185
 
-    def test_no_uris(self):
+    # updated for #1986 (file:// URIs)
+    def test_file_uris_only(self):
         runner = InlineMRJobRunner()
 
         # sanity check
         foo_path = self.makefile('foo')
-        bar_path = os.path.join(self.tmp_dir, 'bar')
+        bar_path = join(self.tmp_dir, 'bar')
         self.assertTrue(runner.fs.exists(foo_path))
-        self.assertFalse(runner.fs.exists(bar_path))
+        self.assertFalse(runner.fs.exists('file://' + bar_path))
 
-        # URI should raise IOError, not return False
+        # non-file:/// URI should raise IOError, not return False
         self.assertRaises(IOError,
                           runner.fs.exists, 's3://walrus/fish')
-        # and it's because we wrapped the local fs in CompositeFilesystem
-        self.assertFalse(runner.fs.local.exists('s3://walrus/fish'))
+
+
+class FileURIsAsInputTestCase(SandboxedTestCase):
+    # regression test for #1986
+
+    def test_file_uris_as_input(self):
+        input1 = self.makefile('input1.txt', b'cat rat bat')
+        input2 = 'file://' + self.makefile('input2.txt', b'dog dog dog')
+
+        job = MRWordFreqCount([input1, input2])
+        job.sandbox()
+
+        with job.make_runner() as runner:
+            runner.run()
+
+            self.assertEqual(dict(job.parse_output(runner.cat_output())),
+                             dict(bat=1, cat=1, dog=3, rat=1))
+
+
+class FileUploadTestCase(SandboxedTestCase):
+
+    # SetupTestCase in test_bin.py covers a lot of this. added this test
+    # so we can ensure that sim runners correctly handle file URIs.
+    # Should probably make this more comprehensive, see #2114
+
+    def test_file_uris(self):
+        f1_path = self.makefile('f1', b'contents')
+        f2_uri = 'file://' + self.makefile('f2', b'stuff')
+
+        job = MROSWalkJob(['--files', '%s,%s' % (f1_path, f2_uri)])
+        job.sandbox()
+
+        with job.make_runner() as runner:
+            runner.run()
+
+            path_to_size = dict(job.parse_output(runner.cat_output()))
+
+            self.assertEqual(path_to_size.get('./f1'), 8)
+            self.assertEqual(path_to_size.get('./f2'), 5)
+
+    def test_archive_uris(self):
+        qux_dir = self.makedirs('qux')
+        self.makefile(join(qux_dir, 'bar'), b'baz')
+
+        qux_tar_gz = make_archive(join(self.tmp_dir, 'qux'), 'gztar', qux_dir)
+        qux_tar_gz_uri = 'file://' + qux_tar_gz
+
+        job = MROSWalkJob(
+            ['--archives', '%s#qux,%s#qux2' % (qux_tar_gz, qux_tar_gz_uri)])
+        job.sandbox()
+
+        with job.make_runner() as runner:
+            runner.run()
+
+            path_to_size = dict(job.parse_output(runner.cat_output()))
+
+            self.assertEqual(path_to_size.get('./qux/bar'), 3)
+            self.assertEqual(path_to_size.get('./qux2/bar'), 3)